+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * The `clean_slate` context for `RedisService` that ensures the service is in an empty state before and after the context.
 * `RedisService`'s new `set_state` method to easily set the state of the internal redis DB.
 * automatic linting on every PR
+* all extra client methods also accept `**kwargs` forwarded to client constructor
 ### Fixed
 * Bug where tests failed on linux because linux doesn't have `host.docker.internal`. Linux now uses IP `172.17.0.1` to target host.
 * Bug where services would fail if trying to run a non-pulled image.